+++ Changelog.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added end-to-end tests for command-line interface.
   #199 by @MaxDesiatov and @mattt.
 
+### Fixed
+
+- Fixed public extensions exposing nested code of all access levels.
+  #195 by @Tunous.
+
 ## [1.0.0-beta.5] - 2020-09-29
 
 ### Added